pyproject.toml
[project]
name = "xevent-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-uinput>=1.0.1",
    "python-xlib>=0.33",
]
